made GA + controller use generics, switch fitness from int to double

+++ src/main/java/SimulationRoundResult.java
@@ -1,10 +1,10 @@
 
 public final class SimulationRoundResult {
-	public int maxFitness;
-	public int minFitness;
-	public int avgFitness;
+	public double maxFitness;
+	public double minFitness;
+	public double avgFitness;
 
-	public SimulationRoundResult(int maxFitness, int minFitness, int avgFitness) {
+	public SimulationRoundResult(double maxFitness, double minFitness, double avgFitness) {
 		this.maxFitness = maxFitness;
 		this.minFitness = minFitness;
 		this.avgFitness = avgFitness;
